requirements.txt
numpy>=1.19.5
pandas>=1.3.0
matplotlib>=3.4.3
seaborn>=0.11.2
scikit-learn>=0.24.2
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.3
Pillow>=8.3.1
streamlit>=1.0.0
tqdm>=4.62.2